+++ lib/Config.php
@@ -9,7 +9,7 @@ use ProcessWire\Inputfield;
 /**
  * SearchEngine Config
  *
- * @version 0.6.1
+ * @version 0.7.0
  * @author Teppo Koivula <teppo.koivula@gmail.com>
  * @license Mozilla Public License v2.0 https://mozilla.org/MPL/2.0/
  */
@@ -85,6 +85,20 @@ class Config extends Base {
 
         // Debugger AJAX endpoint
         (new Debugger)->initAJAXAPI();
+
+        // inject scripts
+        foreach (['Core', 'Config'] as $script) {
+            $this->wire('config')->scripts->add(
+                $this->wire('config')->urls->get('SearchEngine') . 'js/' . $script . '.js'
+            );
+        }
+
+        // inject styles
+        foreach (['config'] as $styles) {
+            $this->wire('config')->styles->add(
+                $this->wire('config')->urls->get('SearchEngine') . 'css/' . $styles . '.css'
+            );
+        }
 
         $fields = $this->wire(new InputfieldWrapper());
 
@@ -186,13 +200,63 @@ class Config extends Base {
         $operator->name = 'find_args__operator';
         $operator->label = $this->_('Operator');
         $operator->description = $this->_('Operator used when finding content. See [documentation for operators](https://processwire.com/docs/selectors/#operators) for more details.');
-        $operator->addOptions([
-            '*=' => '[*=] ' . $this->_('Contains the exact word or phrase'),
-            '~=' => '[~=] ' . $this->_('Contains all the words'),
-            '%=' => '[%=] ' . $this->_('Contains the exact word or phrase (using slower SQL LIKE)'),
-        ]);
+        $operator_options = [
+            '*=',
+            '~=',
+            '%=',
+        ];
+        if (version_compare($this->wire('config')->version, '3.0.160') > -1) {
+            // new operators introduced in ProcessWire 3.0.160
+            $operator_options = array_merge($operator_options, [
+                '~*=',
+                '~~=',
+                '~%=',
+                '~+=',
+                '~|=',
+                '~|*=',
+                '~|%=',
+                '~|+=',
+                '*+=',
+                '**=',
+                '**+=',
+                '#=',
+            ]);
+        }
+        foreach ($operator_options as $operator_option) {
+            $operator_option_class = \ProcessWire\Selectors::$selectorTypes[$operator_option] ?? null;
+            if ($operator_option_class) {
+                $operator_option_class = '\ProcessWire\\' . $operator_option_class;
+                $operator->addOption(
+                    $operator_option,
+                    '[' . $operator_option . '] ' . $operator_option_class::getLabel()
+                );
+            }
+        }
+        $operator->notes .= $this->_('More information in the [documentation page for operators](https://processwire.com/docs/selectors/operators/).');
         $operator = $this->maybeUseConfig($operator);
         $finder_settings->add($operator);
+
+        // operator details
+        $operator_details = $this->wire('modules')->get('InputfieldMarkup');
+        $operator_details->value = '<ul id="pwse-operator-details" class="pwse-operator-details" tabindex="-1" data-toggle-label="' . $this->_('Toggle operator details') . '">';
+        $operator_data_array = \ProcessWire\Selectors::getOperators([
+            'getValueType' => 'verbose',
+        ]);
+        $valid_operators = array_keys($operator->options);
+        foreach ($operator_data_array as $operator_data) {
+            if (!in_array($operator_data['operator'], $valid_operators)) continue;
+            $operator_details_active = $operator->value == $operator_data['operator'] ? 'pwse-operator-details__list-item--active' : '';
+            $operator_details->value .= '<li class="pwse-operator-details__list-item ' . $operator_details_active . '">'
+                . '<div class="pwse-operator-details__header">'
+                . '<button class="pwse-operator-details__button uk-button" data-operator="' . $operator_data['operator'] . '">' . $operator_data['operator'] . '</button>'
+                . '<span class="pwse-operator-details__label">' . $operator_data['label'] . '</span>'
+                . '</div>'
+                . '<div class="pwse-operator-details__description">' . $operator_data['description'] . '</div>'
+                . '</li>';
+        }
+        $operator_details->value .= '</ul>'
+            . '<script>document.getElementById("pwse-operator-details").setAttribute("hidden", "");</script>';
+        $finder_settings->add($operator_details);
 
         return $finder_settings;
     }