+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.25.0] - 2020-07-12
+
+### Added
+- Support for new text search selector operators added in ProcessWire 3.0.160.
+- New JavaScript class for configuration screen features (PWSE_Config).
+
 ## [0.24.0] - 2020-07-07
 
 ### Added